requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
scipy==1.16.0
rdkit-pypi==2022.9.5
xgboost==3.0.2
lightgbm==4.6.0
catboost==1.2.8
torch==2.6.0+cu124
torch-geometric==2.6.1
optuna==4.4.0
tqdm==4.67.1
matplotlib==3.10.0
seaborn==0.13.2